fix(fanin): review wave — bound the subtraction, align demand with real drain, decontaminate the fill feed
Review findings actioned (medium /code-review + verifier wave):

- Config fail-loud (armed decay): executed demand x 2 must fit the inner
  +-max_adjust_ppm authority — legal knob pairs could otherwise demand
  more than a railed ratio can deliver, and the subtraction would
  fabricate inverted-sign clock error (a driven limit cycle, worse than
  the pre-fix chase). Publication also clamps the gauge to the lane
  authority (belt-and-braces, the floor-clamp idiom).
- Demand derivation is now ONE function (DecayParams::step_demand_ppm,
  periods-space, truncation included): config validation, the default
  test pin, and the machine all consume it. The old ms-space test pin
  asserted a zero-margin "4x" that the executed 125.33 ppm does not meet;
  prose now says ~4x and the pin asserts the enforced bound.
- demand publishes only while stepping has actually begun (has_stepped):
  the first interval after (re)activation drains nothing, so publishing
  the mean rate there over-subtracted a ~demand-sized pulse at every
  descent start/resume.
- fill feed descent-compensated into ceiling terms (fill + ceiling-held):
  raw fill fed the ladder's fill-variance falsifier the decay's own
  commanded ramp — same contamination class, few-line fix, bit-identical
  whenever decay is idle.
- demand gauge switched to plain AtomicI64 (the commanded_milli_ppm idiom
  two fields away) instead of a third i64-bits-in-u64 site.
- Honest-comment sweep: guard doc no longer overclaims (standing >400 ppm
  genuine offset freezes decay by design; edges can still cross), decay
  docstring scopes the un-decontaminated refill window, store-count fixed,
  rationale compressed to two owning sites (build_obs, demand_ppm) plus
  pointers, jasper-host-clock's two "resampler idle" doc sites carry the
  #3466 caveat, ADR-0208 records the observable amendment to ADR-0109.
- jasper-control audio_health whitelist carries clamp_count /
  anti_windup_count (the #3464 counters were absent from the curated view).

+++ rust/jasper-fanin/src/state.rs
@@ -889,9 +889,8 @@ impl StateServer {
                 buf.push(',');
                 // Post-lock cushion-decay state (all inert while decay is off):
                 // enabled = startup config; active = actively decaying;
-                // demand_ppm = the live rate demand the descent exerts on the
-                // inner resampler (0 unless active — the exact term subtracted
-                // from the host-clock observable, #3466);
+                // demand_ppm = the live drain demand (the decontamination term
+                // subtracted from the host-clock observable, #3466);
                 // floor = the configured decay floor;
                 // frozen_reason = why decay is paused ("" while actively decaying,
                 // else unlocked / not_l0 / cascade / warmup / at_floor).
@@ -900,8 +899,7 @@ impl StateServer {
                 buf.push(',');
                 push_kv_bool(buf, "active", r.decay_active.load(Ordering::Relaxed));
                 buf.push(',');
-                let demand_ppm =
-                    (r.decay_demand_milli_ppm.load(Ordering::Relaxed) as i64) as f64 / 1000.0;
+                let demand_ppm = r.decay_demand_milli_ppm.load(Ordering::Relaxed) as f64 / 1000.0;
                 push_kv_f64(buf, "demand_ppm", demand_ppm, 2);
                 buf.push(',');
                 push_kv_u64(buf, "floor_frames", r.decay_floor_frames);
@@ -1328,6 +1326,7 @@ fn push_kv_f64_opt(buf: &mut String, key: &str, value: Option<f64>, decimals: us
 #[cfg(test)]
 mod tests {
     use super::*;
+    use std::sync::atomic::AtomicI64;
 
     #[test]
     fn listener_poll_wakes_on_connection_before_shutdown_timeout() {
@@ -1448,7 +1447,7 @@ mod tests {
                         decay_active: Arc::new(AtomicBool::new(false)),
                         decay_floor_frames: 0,
                         decay_frozen_reason: Arc::new(AtomicU64::new(0)),
-                        decay_demand_milli_ppm: Arc::new(AtomicU64::new(0)),
+                        decay_demand_milli_ppm: Arc::new(AtomicI64::new(0)),
                     }),
                     // A lane that HAS been trimmed (fixture): 3 trims, 4608 frames
                     // dropped total, no request currently pending.
@@ -1524,7 +1523,7 @@ mod tests {
                         decay_active: Arc::new(AtomicBool::new(true)),
                         decay_floor_frames: 544,
                         decay_frozen_reason: Arc::new(AtomicU64::new(0)),
-                        decay_demand_milli_ppm: Arc::new(AtomicU64::new(125_330)),
+                        decay_demand_milli_ppm: Arc::new(AtomicI64::new(125_330)),
                     }),
                     trim: Arc::new(TrimControl::test_fixture(0, 0, false)),
                     // The USB DIRECT (combo) lane starts unmuted; the mute-path
